@@ -30,6 +30,9 @@ import androidx.annotation.NonNull;
 import androidx.test.core.app.ApplicationProvider;
 
 import com.microsoft.identity.common.adal.internal.AuthenticationConstants;
+import com.microsoft.identity.common.internal.ui.webview.challengehandlers.CertBasedAuthFactory;
+import com.microsoft.identity.common.internal.ui.webview.challengehandlers.DialogHolder;
+import com.microsoft.identity.common.internal.ui.webview.challengehandlers.YubiKitCertBasedAuthManager;
 import com.microsoft.identity.common.java.ui.webview.authorization.IAuthorizationCompletionCallback;
 import com.microsoft.identity.common.java.providers.RawAuthorizationResult;
 
@@ -99,7 +102,11 @@ public class AzureActiveDirectoryWebViewClientTest {
                         return;
                     }
                 },
-                TEST_REDIRECT_URI);
+                TEST_REDIRECT_URI,
+                new CertBasedAuthFactory(
+                        mActivity,
+                        new YubiKitCertBasedAuthManager(mActivity.getApplicationContext()),
+                        new DialogHolder(mActivity)));
     }
 
     @Test(expected = IllegalArgumentException.class)
