@@ -35,54 +35,32 @@ import java.security.SignatureException;
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
 
+import lombok.AllArgsConstructor;
 import lombok.NonNull;
+import lombok.experimental.Accessors;
 
 /**
- * A basic Signer class.
+ * A very basic Signer class.
  */
-public class BasicSigner implements ISigner{
+@AllArgsConstructor
+@Accessors(prefix = "m")
+public class BasicSigner implements ISigner {
 
-    private static final String TAG = BasicSigner.class.getSimpleName();
+    private final ICryptoFactory mCryptoFactory;
 
     @Override
     public byte[] sign(@NonNull final PrivateKey key,
                        @NonNull final String signingAlgorithm,
                        final byte[] dataToBeSigned) throws ClientException {
-        final Signature signer;
         try {
-            signer = Signature.getInstance(signingAlgorithm);
+            final Signature signer = mCryptoFactory.getSignature(signingAlgorithm);
             signer.initSign(key);
             signer.update(dataToBeSigned);
             return signer.sign();
-        } catch (final NoSuchAlgorithmException e) {
-            throw new ClientException(ClientException.NO_SUCH_ALGORITHM, e.getMessage(), e);
         } catch (final SignatureException e) {
             throw new ClientException(ClientException.SIGNING_FAILURE, e.getMessage(), e);
         } catch (final InvalidKeyException e) {
             throw new ClientException(ClientException.INVALID_KEY, e.getMessage(), e);
-        }
-    }
-
-    public byte[] signWithHMac(final byte[] keyData,
-                               @NonNull final String hmacAlgorithm,
-                               final byte[] dataToBeSigned) throws ClientException {
-        final String methodName = "signWithDerivedKey";
-        try {
-            final Mac sha256HMAC = Mac.getInstance(hmacAlgorithm);
-            final SecretKeySpec secretKey = new SecretKeySpec(keyData, hmacAlgorithm);
-            sha256HMAC.init(secretKey);
-            return sha256HMAC.doFinal(dataToBeSigned);
-        } catch (final NoSuchAlgorithmException e) {
-            final String errorString = hmacAlgorithm + " algorithm does not exist " + e.getMessage();
-            Logger.error(TAG + methodName, errorString, e);
-            throw new ClientException(ClientException.NO_SUCH_ALGORITHM, errorString, e);
-        } catch (final IllegalStateException e) {
-            Logger.error(TAG + methodName, e.getMessage(), e);
-            throw new ClientException(ErrorStrings.ENCRYPTION_ERROR, e.getMessage(), e);
-        } catch (final InvalidKeyException e) {
-            final String errorString = "Key is invalid for signing " + e.getMessage();
-            Logger.error(TAG + methodName, errorString, e);
-            throw new ClientException(ClientException.INVALID_KEY, errorString, e);
         }
     }
 }
